feat(backend): refactored admin
Refactored the admin code into separate modules as it was getting kind
of large w/ lots of unrelated sections in the single module.
