+++ astrosat/admin/admin_settings.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from astrosat.models import AstrosatSettings
+
+
+@admin.register(AstrosatSettings)
+class AstrosatSettingsAdmin(admin.ModelAdmin):
+    pass
